src/test/java/com/chrisdjames1/temperatureanalysis/service/DemoExcelServiceTest.java
package com.chrisdjames1.temperatureanalysis.service;

import org.junit.Before;
import org.junit.Test;

import java.io.File;
import java.io.IOException;
import java.util.List;
import java.util.Map;

import static org.junit.Assert.assertEquals;

public class DemoExcelServiceTest {
    private DemoExcelService demoExcelService;
    private static final String FILE_NAME = "temp.xlsx";
    private String fileLocation;

    @Before
    public void generateExcelFile() throws IOException {
        File currDir = new File(".");
        String path = currDir.getAbsolutePath();
        fileLocation = path.substring(0, path.length() - 1) + FILE_NAME;

        demoExcelService = new DemoExcelService();
        demoExcelService.createDemoPersonsXlsx();
    }

    @Test
    public void whenParsingPOIExcelFile_thenCorrect() throws IOException {
        Map<Integer, List<String>> data
                = demoExcelService.readExcel(fileLocation);

        assertEquals("Name", data.get(0).get(0));
        assertEquals("Age", data.get(0).get(1));

        assertEquals("John Smith", data.get(1).get(0));
        assertEquals("20.0", data.get(1).get(1));
    }
}